# implement AwsRequestProxy that re-uses HTTP client connections