--- a/localstack/aws/forwarder.py
+++ b/localstack/aws/forwarder.py
@@ -2,7 +2,7 @@
 This module contains utilities to call a backend (e.g., an external service process like
 DynamoDBLocal) from a service provider.
 """
-from typing import Any, Callable, Mapping, Optional
+from typing import Any, Callable, Mapping, Optional, Union
 from urllib.parse import urlsplit
 
 from botocore.awsrequest import AWSPreparedRequest, prepare_request_dict
@@ -22,8 +22,83 @@ from localstack.aws.connect import connect_to
 from localstack.aws.skeleton import DispatchTable, create_dispatch_table
 from localstack.aws.spec import load_service
 from localstack.http import Response
-from localstack.http.proxy import forward
+from localstack.http.proxy import Proxy
 from localstack.utils.strings import to_str
+
+
+class AwsRequestProxy:
+    """
+    Implements the ``ServiceRequestHandler`` protocol to forward AWS requests to a backend. It is stateful and uses a
+    ``Proxy`` instance for re-using client connections to the backend.
+    """
+
+    def __init__(
+        self,
+        endpoint_url: str,
+        parse_response: bool = True,
+        include_response_metadata: bool = False,
+    ):
+        """
+        Create a new AwsRequestProxy. ``parse_response`` control the return behavior of ``forward``. If
+        ``parse_response`` is set, then ``forward`` parses the HTTP response from the backend and returns a
+        ``ServiceResponse``, otherwise it returns the raw HTTP ``Response`` object.
+
+        :param endpoint_url: the backend to proxy the requests to, used as ``forward_base_url`` for the ``Proxy``.
+        :param parse_response: whether to parse the response before returning it
+        :param include_response_metadata: include AWS response metadata, only used with ``parse_response=True``
+        """
+        self.endpoint_url = endpoint_url
+        self.parse_response = parse_response
+        self.include_response_metadata = include_response_metadata
+        self.proxy = Proxy(forward_base_url=endpoint_url)
+
+    def __call__(
+        self,
+        context: RequestContext,
+        service_request: ServiceRequest = None,
+    ) -> Optional[Union[ServiceResponse, Response]]:
+        """Method to satisfy the ``ServiceRequestHandler`` protocol."""
+        return self.forward(context, service_request)
+
+    def forward(
+        self,
+        context: RequestContext,
+        service_request: ServiceRequest = None,
+    ) -> Optional[Union[ServiceResponse, Response]]:
+        """
+        Forwards the given request to the backend configured by ``endpoint_url``.
+
+        :param context: the original request context of the incoming request
+        :param service_request: optionally a new service
+        :return:
+        """
+        if service_request is not None:
+            # if a service request is passed then we need to create a new request context
+            context = self.new_request_context(context, service_request)
+
+        http_response = self.proxy.forward(context.request, forward_path=context.request.path)
+        if not self.parse_response:
+            return http_response
+        parsed_response = parse_response(
+            context.operation, http_response, self.include_response_metadata
+        )
+        raise_service_exception(http_response, parsed_response)
+        return parsed_response
+
+    def new_request_context(self, original: RequestContext, service_request: ServiceRequest):
+        context = create_aws_request_context(
+            service_name=original.service.service_name,
+            action=original.operation.name,
+            parameters=service_request,
+            region=original.region,
+        )
+        # update the newly created context with non-payload specific request headers (the payload can differ from
+        # the original request, f.e. it could be JSON encoded now while the initial request was CBOR encoded)
+        headers = Headers(original.request.headers)
+        headers.pop("Content-Type", None)
+        headers.pop("Content-Length", None)
+        context.request.headers.update(headers)
+        return context
 
 
 def ForwardingFallbackDispatcher(
@@ -74,33 +149,21 @@ def HttpFallbackDispatcher(provider: object, forward_url_getter: Callable[[], st
 
 
 def get_request_forwarder_http(forward_url_getter: Callable[[], str]) -> ServiceRequestHandler:
-    def _forward_request(context, service_request: ServiceRequest = None) -> ServiceResponse:
-        if service_request is not None:
-            local_context = create_aws_request_context(
-                service_name=context.service.service_name,
-                action=context.operation.name,
-                parameters=service_request,
-                region=context.region,
-            )
-            # update the newly created context with non-payload specific request headers (the payload can differ from
-            # the original request, f.e. it could be JSON encoded now while the initial request was CBOR encoded)
-            headers = Headers(context.request.headers)
-            headers.pop("Content-Type", None)
-            headers.pop("Content-Length", None)
-            local_context.request.headers.update(headers)
-            context = local_context
-        return forward_request(context, forward_url_getter)
+    """
+    Returns a ServiceRequestHandler that creates for each invocation a new AwsRequestProxy with the result of
+    forward_url_getter. Note that this is an inefficient method of proxying, since for every call a new client
+    connection has to be established. Try to instead use static forward URL values and use ``AwsRequestProxy`` directly.
+
+    :param forward_url_getter: a factory method for returning forward base urls for the proxy
+    :return: a ServiceRequestHandler acting as a proxy
+    """
+
+    def _forward_request(
+        context: RequestContext, service_request: ServiceRequest = None
+    ) -> ServiceResponse:
+        return AwsRequestProxy(forward_url_getter()).forward(context, service_request)
 
     return _forward_request
-
-
-def forward_request(
-    context: RequestContext, forward_url_getter: Callable[[], str]
-) -> ServiceResponse:
-    def _call_http_backend(context: RequestContext) -> Response:
-        return forward(context.request, forward_url_getter())
-
-    return dispatch_to_backend(context, _call_http_backend)
 
 
 def dispatch_to_backend(