… localstack/services/redshift/resource_providers/aws_redshift_cluster.py