<!-- Please refer to the contribution guidelines before raising a PR: https://github.com/localstack/localstack/blob/master/CONTRIBUTING.md -->

<!-- Why am I raising this PR? Add context such as related issues, PRs, or documentation. -->
## Motivation

This PR implements one of the optimizations listed in [this comment](https://github.com/localstack/localstack/issues/1205#issuecomment-1223999205) for improved dynamodb performance.
Previously we were creating a new `Proxy` instance for every `forward_request` call, and therefore never re-using HTTP client connections, which is bad for throughput.

This PR introduces the `AwsRequestProxy` which is a higher-level proxy that satisfies the `ServiceRequestHandler` protocol, and builds on a `Proxy` instance.

This also addresses some comments originally raised in #5796 such as [this one](https://github.com/localstack/localstack/pull/5796#discussion_r842067952).

<!-- What notable changes does this PR make? -->
## Changes

* Introduce `AwsRequestProxy` as high-level proxy for ASF requests to AWS backends
* Change DynamoDB to use one proxy instance per `DynamoDBServer` instance, rather than instantiating one on every request.

## Some preliminary results

I ran a normal `PutItem` loop from [this benchmark](https://github.com/localstack/localstack/issues/1205#issuecomment-1223999205) with 200 calls. I ran each experiment several time and took the median result. I also tested the changes with #9410.

| description | result |
| --- | --- |
| without connection re-use | 200 requests took 3.4139s = 17.0693 ms/op = 58.58 req/sec |
| with connection re-use | 200 requests took 3.0002s = 15.0009 ms/op = 66.66 req/sec |
| **With https://github.com/localstack/localstack/pull/9410** | | 
| without connection re-use | 200 requests took 2.3314s = 11.6570 ms/op = 85.79 req/sec |
| with connection re-use | 200 requests took 2.1307s = 10.6533 ms/op = 93.87 req/sec |

So it adds another 10% performance improvement on top of Ben's optimization, at least for the simple `PutItem` case.

<!-- The following sections are optional, but can be useful! 

## Testing

Description of how to test the changes

## TODO

What's left to do:

- [ ] ...
- [ ] ...

-->